app/requirements.txt
streamlit==1.44.0
mysql-connector-python==9.2.0
pandas==2.2.1
langchain-core==0.3.51
langchain-ollama==0.3.1
langchain==0.3.23
python-dotenv==1.1.0
pyspellchecker==0.7.1  
streamlit-aggrid==0.3.4.post3  